Chane a type from int to int16_t for a little better packing in a structure

+++ include/aio.h
@@ -125,9 +125,9 @@ struct aiocb
   FAR volatile void *aio_buf;    /* Location of buffer */
   off_t aio_offset;              /* File offset */
   size_t aio_nbytes;             /* Length of transfer */
-  int aio_fildes;                /* File descriptor */
-  int8_t aio_reqprio;            /* Request priority offset (not used) */
-  uint8_t aio_lio_opcode;        /* Operation to be performed */
+  int16_t aio_fildes;            /* File descriptor (should be int) */
+  int8_t aio_reqprio;            /* Request priority offset (not used, should be int) */
+  uint8_t aio_lio_opcode;        /* Operation to be performed (should be int) */
 
   /* Non-standard, implementation-dependent data.  For portability reasons,
    * application code should never reference these elements.
